requirements.txt
Automat==20.2.0
Django==3.0
Faker==4.1.0
Jinja2==2.11.2
Markdown==3.1.1
MarkupSafe==1.1.1
Pillow==6.2.0
Pillow-PIL==0.1.dev0
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.9.3
PyYAML==5.1
Pygments==2.6.1
South==1.0.2
Twisted==20.3.0
asgiref==3.2.10
astroid==2.1.0
attrs==20.2.0
autopep8==1.4.3
backports.csv==1.0.7
beautifulsoup4==4.9.1
braintree==4.2.0
certifi==2019.3.9
cffi==1.14.0
chardet==3.0.4
cipher==0.1
colorama==0.4.1
config==0.5.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
covid==2.3.0
covid-india==1.1.1
cryptography==2.9.2
cx-Oracle==7.3.0
cycler==0.10.0
decouple==0.0.7
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
django-allauth==0.39.1
django-annoying==0.10.6
django-bootstrap-datepicker-plus==3.0.5
django-ckeditor==5.8.0
django-common-helpers==0.9.2
django-cors-headers==3.2.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-extra-fields==3.0.2
django-filter==2.1.0
django-guardian==2.2.0
django-heroku==0.3.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-jsonfield==1.4.0
django-mailgun-mime==0.1.7
django-material==1.5.6
django-material-admin==0.0.7
django-materialize==1.0.1a2
django-materializecss-form==1.1.14
django-model-utils==4.0.0
django-multiselectfield==0.1.11
django-payments==0.13.0
django-payments-razorpay==1.0
django-paypal==1.0.0
django-rest-auth==0.9.3
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-tinymce4-lite==1.7.4
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djoser==2.0.5
docopt==0.4.0
et-xmlfile==1.0.1
fcm-django==0.3.4
gunicorn==20.0.4
httpie==2.0.0
hyperlink==20.0.1
idna==2.8
importlib-metadata==1.7.0
incremental==17.5.0
isort==4.3.17
itypes==1.2.0
jdcal==1.4.1
jsmin==2.2.2
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
lxml==4.5.2
mailchimp==2.0.9
matplotlib==3.1.3
mccabe==0.6.1
more-itertools==8.0.2
mysql==0.0.2
mysql-connector-python==8.0.18
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.0.1
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.6.2
ordereddict==1.1
pandas==1.0.1
path==13.1.0
pep8==1.7.1
pip==9.0.3
psycopg2==2.8.4
pyOpenSSL==19.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.9.8
pydantic==1.6.1
pyfcm==1.4.7
pygame==1.9.6
pylint==2.2.2
pylokit==0.8.1
pyotp==2.4.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.3
razorpay==1.2.0
requests==2.21.0
requests-oauthlib==1.2.0
schedule==0.6.0
service-identity==18.1.0
setuptools==39.0.1
simplejson==3.17.0
six==1.12.0
soupsieve==2.0.1
sqlparse==0.3.0
stripe==2.27.0
suds-jurko==0.6
tablib==0.13.0
text-unidecode==1.3
treq==20.9.0
txFCM==0.0.1
typed-ast==1.3.4
uritemplate==3.0.1
urllib3==1.24.2
whitenoise==5.1.0
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
